refactor: drain services/ into application/services/ and config/
authService.ts moves to application/services/ (gains the
"server-only" marker matching its siblings there); api.ts's
BASE_URL moves to config/, alongside the rest of the app's
validated config. apiResponse.ts is deleted outright - it had
zero callers, already superseded by lib/http/server.ts's
httpErrorResponse.
