+++ src/application/services/authService.ts
@@ -1,3 +1,5 @@
+import "server-only";
+
 import { cookies } from "next/headers";
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
